+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## 2.5.3
+ * Feature: Import scale title from .scl files
+ * Bug fix: Fix handling of trailing comments when importing .scl files [#706](https://github.com/xenharmonic-devs/scale-workshop/issues/706)
+
 ## 2.5.2
  * Bug fix: All characters of invalid lines are now encoded in the URL [#688](https://github.com/xenharmonic-devs/scale-workshop/issues/688)
  * Bug fix: Entering 1/1 for the period in "Rank-2 temperament" no longer crashes the application [#690](https://github.com/xenharmonic-devs/scale-workshop/issues/690)